pyproject.toml
[tool.poetry]
name = "jordito"
version = "0.1.0"
description = ""
authors = ["Davies <daviesaraujo@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cryptography = "^40.0.1"
pyrogram = "^2.0.103"
tgcrypto = "^1.2.5"
requests = "^2.28.2"
redis = "^5.2.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
